[BUGFIX] Remove ResourceFactory in CloudinaryDriver
Using the ResourceFactory to fetch a storage object leads to infinite
nesting of method calls as the initialization process of the storage calls
the `hasFolder()` method of the driver which in turn starts an
initialization of the storage.

This change removes usage of ResourceFactory and replaces the storage
object with plain storageConfiguration or storageId values.