[BUGFIX] Use identifier_hash in CloudinaryScanService
The `fileExistsInStorage()` method currently performs a SQL query to count
records, but suffers from slow performance due to the absence of an index on
the `identifier` field. This results in a full table scan.

To address this issue, this commit refactors the method to utilize the
pre-existing `identifier_hash` field, specifically designed for efficient data
searches.